+++ R/rAMMI.R
@@ -45,6 +45,7 @@
 #'
 #' library(geneticae)
 #' # Data without replication
+#' library(agridat)
 #' data(yan.winterwheat)
 #' BIP_AMMI <- rAMMI(yan.winterwheat, genotype = "gen", environment = "env",
 #'                   response = "yield", type = "AMMI")