+++ NEWS.md
@@ -2,7 +2,8 @@
 
 This is a development version
 
-Changed default option to "symmetrical" in the SVP argument of the GGEmodel() function.
+Changed default option to "symmetrical" in the SVP argument of the GGEmodel()
+function.
 
 # geneticae 0.1.0
 
@@ -11,3 +12,7 @@ First submission to CRAN.
 # geneticae 0.0.9000
 
 This is the first development version of the package.
+
+
+
+