+++ NEWS.md
@@ -1,6 +1,4 @@
-# geneticae 0.1.9000
-
-This is a development version
+# geneticae 0.2.0
 
 Changed default option to "symmetrical" in the SVP argument of the GGEmodel()
 function.